repalce `id` `SERIAL` type with `INTEGER` I made an incorrect assumption about how `SERIAL` worked under-the-hood. I assumed that postgres checked for the next value in the series if not explicitly defining the value for `id` when inserting a row. What actually is happening, is postgres keeps a separate record of the next `id` series value, and updates it whenever it inserts a new row.
This didn't matter, because allowing dynamic definitions of these types of entities directly contradicts my design and implementation of static game data and it's seeding into the database.

It is possible that INTEGER ends up being too small, but this project shouldn't grow large enough to make that an issue.